backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
aiosqlite>=0.20.0
pydantic>=2.0
pydantic-settings>=2.0
python-nmap>=0.7.1
aiohttp>=3.9
requests>=2.31
paramiko>=3.4
beautifulsoup4>=4.12
cryptography>=42.0
dnspython>=2.6
jinja2>=3.1
weasyprint>=62.0
python-multipart>=0.0.9
websockets>=12.0
scapy>=2.5
PyJWT>=2.8
httpx>=0.27
